requirements.txt
py_getch==1.0.1
Requests==2.28.2
setuptools==47.1.0
pyinstaller==5.10.1
